延迟精度调整

+++ src/main/java/com/feeyo/kafka/config/KafkaPoolCfg.java
@@ -27,7 +27,7 @@ public class KafkaPoolCfg extends PoolCfg {
 	// topicName -> topicCfg
 	private volatile Map<String, TopicCfg> topicCfgMap = null;
 
-	public KafkaPoolCfg(int id, String name, int type, int minCon, int maxCon, int maxLatencyThreshold) {
+	public KafkaPoolCfg(int id, String name, int type, int minCon, int maxCon, float maxLatencyThreshold) {
 		super(id, name, type, minCon, maxCon, maxLatencyThreshold);
 	}
